SteemHunt Privacy Policy
A Frequently Asked Question section (FAQs) is without a doubt the most commonly used tool in self-service. Service users tend to feel that their question is the most important question, and they are right. With Frequently Asked Questions, new service users  (in this case, non steemit.com user) will find answers to possible questions with ease and in a very short time frame.